Add barrier to sync file i/o
We have had trouble with MPI Messages that do not arrive:
The slave processes (rank != 0) quit already while the master is still
writing e.g. netcdfs. This leads to a deadlock in MPI_Recv..
(Slaves were stuck in zmq-exit, master in MPI_Recv) so the barrier
needed to be inserted to ensure that all messages are handled correctly.