Implement easier API for melissa_init
Some parameters of melissa_init can be left. Further I think it is
cleaner to put the simu_id into an environment variable as so the user
is not required to add it as an additional parameter to his simulation
code. On the other handside this requires changes in the options.py to
set the MELISSA_SIMU_ID environment variable correctly.