Launcher: fix heterogeneous jobs with Slurm
The code for starting heterogeneous jobs with Slurm was broken and I put
the blame squarely on Slurm and the Slurm documentation. The necessity
to pass the the argument `--pack-group=<expr>` to srun is mentioned in
only one place: in the one-paragraph documentation of this parameter.

Thanks to a presentation called "Batch Usage on JSC. Introduction to
Slurm" by Chrysolovantis Paschoulas (Forschungszentrum Jülich) I was
able to figure out the correct call.