CI: move Dockerfiles to melissa/melissa-test repo
The reason to move the Dockerfiles to melissa/melissa-test> is the fact
that are also used for Melissa DA and Melissa DA needs additional
packages.

The occasion to do it is that I accidentally built a Docker image with
the Dockerfiles in this repository and broke the Melissa DA CI (see,
e.g., Job 904390).