CMake: fix Python tests on RedHat Enterprise Linux
On the Jean Zay supercomputer, all Python tests were failing because the
unittest module was unable to import the modules it was supposed to
test. In particular, the module seemed unable to tell apart the
filesystem path `build/build.melissa-sa` from a Python package named
`build` containing a package `melissa-sa`.

The exact cause of the problem is not clear to me since these tests
succeed inside the CentOS 7 and 8 LXD containers.

* RedHat Enterprise Linux 8.1
* Python 3.6.9

```
1: Test command: /usr/bin/python3 "-m" "unittest" "/linkhome/rech/genini01/use61mq/build/build.melissa-sa/melissa/launcher/job_management.py"
1: Test timeout computed to be: 1500
1: E
1: ======================================================================
1: ERROR: /linkhome/rech/genini01/use61mq/build/build (unittest.loader._FailedTest)
1: ----------------------------------------------------------------------
1: ImportError: Failed to import test module: /linkhome/rech/genini01/use61mq/build/build
1: Traceback (most recent call last):
1:   File "/usr/lib64/python3.6/unittest/loader.py", line 153, in loadTestsFromName
1:     module = __import__(module_name)
1: ModuleNotFoundError: No module named '/linkhome/rech/genini01/use61mq/build/build'
```