Move regression test into build-and-test
Move the regression test for issue #105 (MPI C++ bindings broken by a
Melissa header) from the Melissa tests into the find-and-link test for
the Melissa C++ bindings. Moreover, the test is modified and uses only
the MPI C bindings because the C++ MPI bindings were deprecated in the
MPI 2.2 standard released in 2009 and may not be present at all.

https://github.com/mpi-forum/mpi-forum-historic/issues/150