Remove code enabled by CHECK_SIMU_DECONNECTION
The CMake option `CHECK_SIMU_DECONNECTION` was removed by merge request
!29 because it was unused by CMake but I did not check if there is C or
Fortran code enabled by this variable. This commit removes the code that
could be enabled by defining a pre-processor flag called
`CHECK_SIMU_DECONNECTION`.

The purpose of the code was to allow a simulation to tell a server when
it was about to terminate. Just for this purpose it would create another
pair of ZeroMQ sockets because the ZeroMQ API does not ever inform
callers about disconnects.